updated patientcode format

+++ database/migrations/2017_02_11_070904_create_patients_table.php
@@ -36,7 +36,7 @@ class CreatePatientsTable extends Migration
             $table->text('allergies');
             $table->string('bloodgroup',10);
             $table->string('idproof');
-            $table->smallInteger('patientcode')->unsigned()->unique();
+            $table->integer('patientcode')->unsigned()->unique();
             $table->integer('created_by')->unsigned();
             $table->foreign('created_by')->references('id')->on('users');
             //$table->date('dob');
